requirements.txt
# Mínimos para retriever MMR + Chroma alinhados ao LangChain 0.3.x
streamlit>=1.28.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-chroma>=0.1.0
langchain-groq
langchain-huggingface
sentence-transformers
python-dotenv
protobuf<5.0.0
